pyproject.toml
[project]
name = "4-etl-to-aws"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.1"
dependencies = [
    "boto3>=1.36.21",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
